smartjobportal/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
PyMuPDF==1.24.2
PyPDF2==3.0.1
pdfminer.six==20221105
spacy==3.7.4
sentence-transformers==2.7.0
Pillow==10.3.0
